Refactor lattice_search sub-module into sub-package
